Lots of changes. See full log entry.
Since Rick started a trend for shorter filenames, I've done some
rearranging. 'elcap' is being phased out as a prefix in these files
(though not yet excised completely). His dd-background.csv has been
changed to background-dd.csv so that all the background-form resources
appear "together" in a directory listing, but he may have intended for
all the data dictionaries to appear together that way; not sure
yet. It can be changed easily enough if that ws his intent.

Added the meetings page and the Rick/Domenic Q&A pages, removed the
'-dd2' stuff for now (it was essentially a list of suggested changes
to the data dictionary, but it mostly consisted of asking that all
T/F|Y/N radio buttons be replaced with checkboxes, and making the
corrections that are in the '-ddr' file.

I'm not sure what to do with Rick's Mumpsified data dictionary, which
contains less information than is in the formal one, so I just made it
downloadable from a link and gave it the tiniest bit of documentation.

Started rearranging the elcap-forms index page, but that's not done
yet. I guess I need to remove the redundant information from the main
project page while I'm at it, if it's all going to live down here (and
George has already passed the direct link rather than the link to the
project page, so yeah, better do that).

(I"m not sure what it is about Mumps programmers, but they seem to
prefer a fragmented information model. I'm still going to try to
maintain a useful project page that points to all the necessary
resources.)

I'm also trying to snag important notices to put them on the project
page (like Drew's admonition not to freely talk about VA{nbhy}PALS).

Oh, and I also added a link to Laura Cartwright's VA{nbhy}PALS Dropbox
folder. I wish Dropbox would handle multiple accounts more easily than
it currently does.

I'm still fiddling with the info-box elements; it turns out that
browser support for external SVGs is still not living up to the dang
promise. I may have to make these internal SVGs in JavaScript; it's
technically "not hard," but ugh.