VAP-165 - Converted modal validation message to native form validation
This change replaced the original modal validation message for the CT Evaluation form to a native formvalidation message. This is a custom validator that looks at a single radio button, and when it's value is "y", ensures that at least one other radio button in a given container is also "y".