VAP-207 : updated intake form title
Note that this only updated the form title, without adding any
additional markup.  The page title is specified in python and provided
to the page. It didn't seem worthwhile to either change all forms (too
much) or to make an exception for this form (where the title wouldn't be
escaped to allow markup)