VAP-3: updated HTML to be roughly 508 compliant
With an internal system like this, it's tricky to run external tools against it for accessibility. The POST URL also makes it difficult to point it to any of the forms. The accessibility analyzer I ran was one built into my IDE (IntellJ). The changes herein resolve all issues reported by that inspection.

 - Added labels (albeit hidden, which is acceptable) to all form inputs
 - Updated image scaling to use CSS over width property