VAP-3 section 508 accessibility (#57)
* VAP-3: updated HTML to be roughly 508 compliant

With an internal system like this, it's tricky to run external tools against it for accessibility. The POST URL also makes it difficult to point it to any of the forms. The accessibility analyzer I ran was one built into my IDE (IntellJ). The changes herein resolve all issues reported by that inspection.

 - Added labels (albeit hidden, which is acceptable) to all form inputs
 - Updated image scaling to use CSS over width property

* VAP-3: updated per code review.

 - replaced image scaling with pre-scaled images
 - removed lang attributes from <script> elements
 - updated forms macro to supress for attributes on a label when argument for its value is empty

* VAP-3: remove image height from layout