VAP-333 remove label for attr (#76)
* VAP-333 Remove for attributes for specific question selections

This commit updates HTML labels such that when clicked will NOT select a radio/checkbox/select option. This was originally added for accessibility, but was deemed unnecessary for 508 compliance. Labels for questions are now simple labels, not linked to a specific answer. Screen readers would supposedly read the label first, then present the user with the applicable options simply because of the order in the HTML form.

* VAP-333 Regenerated HTML forms