Adding a data dictionary directory to documents.
This contains the TSV files provided by George (with a couple of minor
bugfixes that will eventually have to be addressed), and the
corresponding HTML files generated by the tsv-dd-to-html.py Python
script (included). This should be considered a first-draft form; there
are definitely a few improvements that can be made, and I should make
a better-commented version.