Audit: 'sbmphthv' converted from numeric to text field.
Per Claudia on 2021-05-27, this should be just a text field "so that
the user can put in a complete blood pressure reading
(systolic/diastolic) or a single number." Removed the "numeric
decimalformat" from the class, and removed the relevant validation
lines.

Also commented out the TSV generator in build.sh for the moment; I'm
working on a new one in any case as part of the audit.