Added "Reason for cancellation?" field (bxcr) to Biopsy form.
Added per Artit, 2021-05-18.

The only remaining problem I'm having with this is that I can't figure
out how to have the form disabled when the form loads. It's disabled
when you select Yes or No and enabled when you hit Cancel, but I'd
like it to be disabled when the form loads, and I haven't worked out
how that happens yet.