Adding Dom's requested changes, and a new field.
Dom provided a number of helpful suggestions and caught some errors
I'd made. Also, prompted by a query from Dom, we were able to
determine that a "specify" field should be added to the "Discharge
location?" question, and this was done. However, Artit said it
shouldn't be required, and Linda suggested to not have it be
conditionally enabled, as it had a use whichever checkbox was
selected. The new field (rbrdls) has been added to the form-fields
TSV, and a new Excel file has been generated.