The Horrible Kludge redux. (Longer explanation below.)
As has been already noted, but not yet fixed, there's a routine
somewhere (in Mash?) that truncates lines in the .html files to 255
characters, which is not compliant with the standards (they explicitly
allow unlimited line lengths).

In the script that compiles the .jinja2 files to .html, the resulting
files may have line lengths exceeding that limit. Most don't, but a
few files (register.html and editparticipant.html) wind up with a
couple of 400-character-plus monster lines due to HTML tags with
necessary attributes for the application.

Because of the bug in the aforementioned routine, we've had to
manually edit these two files before a patch is applied.

To find out if any files have lines that exceed 255 characters in
length, you can run the following command in the docs/www directory:

$ for FN in *.html; do \
  L=`cat $FN | awk \
     'BEGIN {maxlen = 0;}
      {ll = length($0); if (ll > maxlen) maxlen = ll;}
      END {print maxlen;}'`; \
  if [ $L -gt 255 ]; then printf "%25s = %d\n" $FN $L; fi; \
  done

If you want to see the line lengths of all the files, you can run this
command in that directory instead:

$ for FN in *.html; do \
  L=`cat $FN | awk \
     'BEGIN {maxlen = 0;}
      {ll = length($0); if (ll > maxlen) maxlen = ll;}
      END {print maxlen;}'`; \
  printf "%25s = %d\n" $FN $L; \
  done

Ultimately, this manual intervention will no longer be necessary when
the bug in the routine is found and addressed properly.