Adding the 'bxcr' field ("Reason for cancellaton?") to biopsy form.
I still have a bug I haven't been able to track down: when the form
loads, the field in question is enabled; it shouldn't be. Other fields
are disabled on load, and I'm not sure what I'm doing wrong;
otherwise, the field works as you expect: if you select Yes or No
under "Was a biopsy performed?", the field is disabled, but if you
select Cancel, the field enables and shows that it's now a required
value.

This needs to go in 18.12, however, and the problem is effectively
cosmetic. I look forward to finding out how I got it wrong.