Change in recipe file for 18-x-15-t2 and all releases going forward.
Specifically, the recipe format moves from 2.2-a2 to 2.2-a3, and a new
header tag is added: @app-page. This gives the root of the website for
the application.

That website root must have a dist/ directory (in our case, it's just
a symbolic link to the application repository's dist/ directory), and
that should contain all the versions (like, 18-x-15) with test
releases under those (t1, t2, etc.).

We really need to document this better---and also need to figure out
where to put that documentation.