Reorganizing docs/ under dist/. See full commit text for reason.
This is part 1 of this reorganization. Part of this website is being
leveraged to provide project and release documentation. Before now,
this has involved both the dist/ and the docs/ directory, but this
proved a little awkward to manage. Since documentation is intended to
be distributed, it makes sense to put docs/ under dist/, which
simplifies a few things. However, there is also a symbolic link so
that the original docs/ can still be used. (This was important because
a number of our distribution processes currently rely on that. We
really need to revisit those at some point, however.)

There will be a few more commits along this line.