SAMI 18-17-t4 release draft 1.
There were a couple of issues with this. This was the first time the
recipe has included an auxfile that was just a specific file rather
than a directory tree; this necessitated a change to dtool to detect
the difference and behave accordingly.

Also, when there are auxfiles involved, the recipe list currently
needs to include the `upgrade-auxfiles` script. I added that line to
the recipe file.