This repository now manages its own web content.
In other words, a symbolic link from the VEN website will link to the
root of the repo, and from there, it'll be a miniwebsite. There's no
style or configuration information here; that's all managed from the
main website---but the content can be changed from here.

Also included a number of pending files and a bit of miscellany.